% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_horizontal_lines}
\alias{add_horizontal_lines}
\title{Add horizontal lines and their legend labels to a plot}
\usage{
add_horizontal_lines(
  plot,
  plot_data,
  agg_label,
  color_comb,
  new_hline_col,
  hline_arb = NULL,
  hline_arb_color = "red",
  hline_arb_label = NULL,
  hline_vars = NULL,
  hline_vars_colors = NULL,
  hline_vars_labels = NULL
)
}
\description{
Add horizontal lines and their legend labels to a plot
}
