% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{caption_loqs_label}
\alias{caption_loqs_label}
\title{Add footnote to identify LLOQ and ULOQ values identified from data}
\usage{
caption_loqs_label(loqs_data)
}
\arguments{
\item{loqs_data}{(`data frame`)\cr loqs_data data set containing assay data with potential LOQ values}
}
\description{
Add footnote to identify LLOQ and ULOQ values identified from data
}
\examples{
library(scda)
ADLB <- synthetic_cdisc_data("latest")$adlb
caption_label <- goshawk:::caption_loqs_label(loqs_data = ADLB)

}
