% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goshawk_core_packages.R
\docType{data}
\name{.mcr}
\alias{.mcr}
\title{loading namespace in non-standard attachment for private function mc.deming}
\format{An object of class \code{environment} of length 122.}
\usage{
.mcr
}
\description{
loading namespace in non-standard attachment for private function mc.deming
}
\keyword{datasets}
