% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_spaghettiplot.R
\name{g_spaghettiplot}
\alias{g_spaghettiplot}
\title{Spaghetti plot}
\usage{
g_spaghettiplot(data, subj_id = "USUBJID", biomarker_var = "PARAMCD",
  biomarker, value_var = "AVAL", unit_var = "AVALU", trt_group,
  trt_group_level = NULL, time, time_level = NULL, ymin = NA, ymax = NA,
  facet_ncol = 2, hline = NULL, rotate_xlab = FALSE)
}
\arguments{
\item{data}{data frame with variables to be summarized and generate statistics which will display in the plot.}

\item{subj_id}{unique subject id variable name.}

\item{biomarker_var}{name of variable containing biomarker names.}

\item{value_var}{name of variable containing biomarker results.}

\item{trt_group}{name of variable representing treatment group.}

\item{trt_group_level}{vector that can be used to define the factor level of trt_group.}

\item{time}{name of vairable containing visit names.}

\item{time_level}{vector that can be used to define the factor level of time.}

\item{ymin}{y-axis lower limit.}

\item{ymax}{y-axis upper limit.}

\item{facet_ncol}{number of facets per row.}

\item{hline}{numeric value represnting intercept of horizontal line.}

\item{biomaker}{biomarker name to be analyzed.}

\item{roate_xlab}{boolean whether to rotate x-axis labels.}
}
\value{
\code{ggplot} object
}
\description{
This function is rendered by teal.goshawk module
}
\details{

}
\examples{

library(dplyr)

ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISIT = paste0("visit ", 1:10),
  ARM = c("ARM A", "ARM B", "ARM C"),
  PARAMCD = c("CRP", "IGG", "IGM")
)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100
ANL$AVALU <- 'mg'

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

g_spaghettiplot(data = ANL,
                subj_id = 'USUBJID',
                biomarker_var = 'PARAMCD',
                biomarker = 'CRP',
                value_var = 'AVAL',
                trt_group = 'ARM',
                time = 'VISIT',
                hline = NULL,
                rotate_xlab = FALSE)
}
\author{
Wenyi Liu (wenyi.liu@roche.com)
}
