% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geom_arb_hline}
\alias{geom_arb_hline}
\alias{geom_arb_vline}
\title{Abitrary lines for `ggplot2`}
\usage{
geom_arb_hline(
  mapping,
  data,
  ...,
  yintercept = numeric(0),
  color = "red",
  label = "Horizontal line",
  show.legend = NA
)

geom_arb_vline(
  xintercept = numeric(0),
  color = "red",
  label = "Horizontal line",
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or \code{\link[ggplot2:aes_]{aes_()}}.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{yintercept}{Parameters that control the
position of the line. If these are set, \code{data}, \code{mapping} and
\code{show.legend} are overridden.}

\item{label}{(`character`)\cr
Label to be rendered in the legend. Should be a single string or vector of length
equal to length of `xintercept`.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{xintercept}{Parameters that control the
position of the line. If these are set, \code{data}, \code{mapping} and
\code{show.legend} are overridden.}
}
\description{
Abitrary lines for `ggplot2`
}
\examples{
library(ggplot2)

data <- data.frame(x = seq_len(10), y = seq_len(10), color = rep(c("a", "b"), each = 5))
ggplot2(data, aes(x = x, y = y, color = color)) +
  geom_point() +
  geom_arb_hline(yintercept = c(2, 5), color = "blue", label = c("h1", "h2"), linetype = 2)
library(ggplot2)

data <- data.frame(x = seq_len(10), y = seq_len(10), color = rep(c("a", "b"), each = 5))
ggplot2(data, aes(x = x, y = y, color = color)) +
  geom_point() +
  geom_arb_vline(xintercept = c(2, 5), color = "blue", label = c("h1", "h2"), linetype = 2)
}
