% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_axes_line.R
\name{geom_straight_lines}
\alias{geom_straight_lines}
\alias{geom_arb_hline}
\alias{geom_arb_vline}
\alias{geom_range_hline}
\alias{geom_range_vline}
\title{Straight lines for `ggplot2`}
\usage{
geom_arb_hline(
  yintercept,
  label = "Horizontal line",
  color = "red",
  legend_title = "Horizontal line(s)",
  ...
)

geom_arb_vline(
  xintercept,
  label = "Vertical line",
  color = "red",
  legend_title = "Vertical line(s)",
  ...
)

geom_range_hline(
  vars,
  data,
  color = "green",
  label = vars,
  legend_title = "Horizontal range line(s)",
  ...
)

geom_range_vline(
  vars,
  data,
  color = "green",
  label = vars,
  legend_title = "Vertical range line(s)",
  ...
)
}
\arguments{
\item{yintercept}{(`numeric`)\cr
Position of the horizontal line(s) on the y-axis}

\item{label}{(`character`)\cr
Label to be rendered in the legend.
Should be a single string or vector of length equal to length of `xintercept`.}

\item{color}{(`character`)\cr
Valid color convertible to RGB scale by [grDevices::col2rgb()].
Should be a single string or vector of length equal to length of `xintercept`.}

\item{xintercept}{(`numeric`)\cr
Position of the vertical line(s) on the x-axis}

\item{vars}{(`character`)\cr
names of variables `(ANR*)` or values `(*LOQ)` identifying intercept values.}
}
\description{
Abitrary lines for `ggplot2`
}
\examples{
# horizontal arbitrary lines
library(ggplot2)

data <- data.frame(x = seq_len(10), y = seq_len(10), color = rep(c("a", "b"), each = 5))
ggplot(data, aes(x = x, y = y, color = color)) +
  geom_point() +
  goshawk:::geom_arb_hline(yintercept = c(2, 5), color = "blue", label = c("h1", "h2"), linetype = 2)

# vertical arbitrary lines
library(ggplot2)

data <- data.frame(x = seq_len(10), y = seq_len(10), color = rep(c("a", "b"), each = 5))
ggplot(data, aes(x = x, y = y, color = color)) +
  geom_point() +
  goshawk:::geom_arb_vline(xintercept = c(2, 5), color = "blue", label = c("h1", "h2"), linetype = 2)

# horizontal range
library(ggplot2)

data <- data.frame(
  x = seq_len(10),
  y = seq_len(10),
  color = rep(c("a", "b"), each = 5),
  lower = rep(c(2, 3), each = 5),
  upper = rep(c(7, 8), each = 5)
)
ggplot(data, aes(x = x, y = y, color = color)) +
   geom_point() +
   goshawk:::geom_range_hline(
    vars = c("lower", "upper"),
    data = data.frame(lower = 2, upper = 7),
    color = "blue",
    linetype = 2
  )


# vertical range
library(ggplot2)

data <- data.frame(
  x = seq_len(10),
  y = seq_len(10),
  color = rep(c("a", "b"), each = 5),
  lower = rep(c(2, 3), each = 5),
  upper = rep(c(7, 8), each = 5)
)
ggplot(data, aes(x = x, y = y, color = color)) +
   geom_point() +
   goshawk:::geom_range_vline(
    vars = c("lower", "upper"),
    data = data.frame(lower = 2, upper = 7),
    color = "blue",
    linetype = 2
  )

}
\keyword{internal}
