% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_variables.R
\name{global_variables}
\alias{global_variables}
\title{To prevent "no visible binding for global variable" note during R CMD check.}
\description{
To prevent "no visible binding for global variable" note during R CMD check.
}
