% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goshawk-package.R
\docType{package}
\name{goshawk-package}
\alias{goshawk}
\alias{goshawk-package}
\title{goshawk: Longitudinal Visualization Functions}
\description{
Functions that plot and summarize biomarkers/labs of interest. Visualizations include: box plot, correlation plot, density distribution, line plot and spaghetti plot. Data are expected in ADaM structure. Requires analysis subject level (ADSL) and analysis laboratory (ADLB) data sets. Beyond core variables, Limit of Quantification flag variable (LOQFL) is expected with levels 'Y', 'N' or NA.
}
\author{
\strong{Maintainer}: Nick Paszty \email{nick.paszty@gene.com}

Authors:
\itemize{
  \item Junlue Zhao \email{zhaoj88@gene.com}
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}
  \item Maciej Nasinski \email{maciej.nasinski@contractors.roche.com}
  \item Konrad Pagacz \email{konrad.pagacz@contractors.roche.com}
  \item Ondrej Slama \email{ondrej.slama@roche.com}
  \item Adrian Waddell \email{adrian.waddell@gene.com}
  \item Wenyi Liu \email{liu.wenyi@roche.com}
  \item Jeffrey Tomlinson \email{tomlinson.jeffrey@roche.com}
  \item Bali Toth \email{toth.balazs@gene.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
