% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goshawk.R
\docType{package}
\name{goshawk}
\alias{goshawk}
\title{The goshawk package provides longitudinal visualizations of lab/biomarker data}
\description{
The goshawk package provides longitudinal visualizations of lab/biomarker data
}
