% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_identify_loq_values}
\alias{h_identify_loq_values}
\title{Helper for identifying any \code{LLOQ} and \code{ULOQ} values in \code{LBSTRESC}. Outcome drives
horizontal line functionality display and legend labeling along with display
of values in footnote.}
\usage{
h_identify_loq_values(loqs_data)
}
\arguments{
\item{loqs_data}{(\verb{data frame})\cr \code{loqs_data} data set containing assay data with potential \code{LOQ} values}
}
\description{
Helper for identifying any \code{LLOQ} and \code{ULOQ} values in \code{LBSTRESC}. Outcome drives
horizontal line functionality display and legend labeling along with display
of values in footnote.
}
\details{
Biomarker Sciences would like to have \code{LLOQ} and \code{ULOQ} values available for
reference in the visualizations. This also aids in setting the data constraint
ranges when \code{goshawk} functions are run from \code{teal.goshawk} \code{UI}.
}
\examples{
library(scda)

ADLB <- synthetic_cdisc_data("latest")$adlb

goshawk:::h_identify_loq_values(loqs_data = ADLB)
}
\keyword{internal}
