% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_axes_line.R
\name{hline_axis_label}
\alias{hline_axis_label}
\title{Add labels to the y axis equaling to the values of the lines}
\usage{
hline_axis_label(plot, replace = FALSE)
}
\arguments{
\item{plot}{('ggplot') a ggplot with either horizontal or vertical lines}

\item{replace}{('logical') flag to determine if the line value labels should replace the entire axis}
}
\value{
\code{ggplot} object
}
\description{
Add labels to the y axis equaling to the values of the lines
}
\examples{

library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  goshawk:::geom_axes_lines(
    hline_arb = c(12, 25, 30),
    hline_arb_color = "red",
    hline_arb_label = "Hori Line"
  )
goshawk:::hline_axis_label(p)
}
\keyword{internal}
