% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_axes_line.R
\name{line_axis_label}
\alias{line_axis_label}
\title{Add labels to the y/x axes equaling to the values of the lines}
\usage{
line_axis_label(plot)
}
\arguments{
\item{plot}{('ggplot') a ggplot with either horizontal or vertical lines}
}
\value{
(\code{ggplot}) with the new axes breaks
}
\description{
Add labels to the y/x axes equaling to the values of the lines
}
\examples{

library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) +
geom_point() +
 goshawk:::geom_axes_lines(
   hline_arb = c(12, 25, 30),
   hline_arb_color = "red",
   hline_arb_label = "Hori Line"
 )
goshawk:::line_axis_label(p)
}
