% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarytable_av.R
\name{t_summarytable_av}
\alias{t_summarytable_av}
\title{Function to create a table of descriptive summary statistics to accompany plots
that present all visit data based on analysis day.}
\usage{
t_summarytable_av(data, trt_group, param_var, param, xaxis_var,
  facet_var = "AVISITCD", loq_flag_var = "LOQFL", ...)
}
\arguments{
\item{data}{name of data frame to summarize.}

\item{trt_group}{treatment group variable name e.g. ARM.}

\item{param_var}{name of variable containing biomarker codes e.g. PARAMCD.}

\item{param}{biomarker to visualize e.g. IGG.}

\item{xaxis_var}{name of variable containing biomarker results displayed on X-axis e.g. AVAL.}

\item{facet_var}{name of variable containing visit values e.g. AVISITCD.}

\item{loq_flag_var}{name of variable containing LOQ flag e.g. LOQFL.}

\item{...}{additional options}
}
\description{
Output descriptive summary statistics table as a data frame.
Includes biomarker, treatment, visit, n, meand, median, sd, min, max,
%missing values, % LOQ values.
}
\details{
provide additional information as needed. link to specification file
\url{http://rstudio.com}
}
\examples{

\dontrun{

# Example using ADaM structure analysis dataset.

library(dplyr)
library(goshawk)
library(random.cdisc.data)
library(stringr)

# original ARM value = dose value
arm_mapping <- list("A: Drug X" = "150mg QD", "B: Placebo" = "Placebo",
"C: Combination" = "Combination")

ASL <- radsl(N = 20, seed = 1)
ALB <- radlb(ASL, visit_format = "WEEK", n_assessments = 7L, seed = 2)
ALB <- ALB \%>\%
mutate(AVISITCD = case_when(
AVISIT == "SCREENING" ~ "SCR",
AVISIT == "BASELINE" ~ "BL", grepl("WEEK", AVISIT) ~ paste("W", trimws(substr(AVISIT, start = 6,
stop = str_locate(AVISIT, "DAY")-1))),
TRUE ~ as.character(NA))) \%>\%
mutate(AVISITCDN = case_when(AVISITCD == "SCR" ~ -2,
AVISITCD == "BL" ~ 0, grepl("W", AVISITCD) ~ as.numeric(gsub("\\\\D+", "", AVISITCD)),
TRUE ~ as.numeric(NA))) \%>\%
# use ARMCD values to order treatment in visualization legend
mutate(TRTORD = ifelse(grepl("C", ARMCD), 1,
ifelse(grepl("B", ARMCD), 2,
ifelse(grepl("A", ARMCD), 3, NA)))) \%>\%
mutate(ARM = as.character(arm_mapping[match(ARM, names(arm_mapping))])) \%>\%
mutate(ARM = factor(ARM) \%>\% reorder(TRTORD))

tbl <- t_summarytable_av(data = ALB,
               trt_group = "ACTARM",
               param_var = "PARAMCD",
               param = c("ALKPH"),
               xaxis_var = "AVAL",
               facet_var = "ACTARM",
               loq_flag_var = "LOQFL")
tbl

}

}
\author{
Nick Paszty (npaszty) paszty.nicholas@gene.com

Balazs Toth (tothb2)  toth.balazs@gene.com
}
