% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_hori_line_args}
\alias{validate_hori_line_args}
\title{validate arbitrary horizontal lines}
\usage{
validate_hori_line_args(
  data,
  hline_arb = NULL,
  hline_arb_color = "red",
  hline_arb_label = NULL,
  hline_vars = NULL,
  hline_vars_colors = NULL,
  hline_vars_labels = NULL
)
}
\description{
validate arbitrary horizontal lines
}
