% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_axes_line.R
\name{validate_line_args}
\alias{validate_line_args}
\title{Validate line arguments given in the parameters against the data.}
\usage{
validate_line_args(data, vars = character(0), color = "green", label = vars)
}
\arguments{
\item{data}{(\code{data.frame})\cr
should contain \code{vars} which will be used to create the plot.}

\item{vars}{(\code{character})\cr
names of variables to take the values from. Only first value from the variable will be
applied.}

\item{color}{(\code{character})\cr
colors for the lines defined by variables.}

\item{label}{(\code{character})\cr
labels for the legend to the lines defined by variables.}
}
\value{
(\code{data.frame}) containing the \code{values},\code{colors} and \code{labels} fields defining attributes
for horizontal or vertical lines.
}
\description{
helper function to be called by \code{\link[=geom_axes_lines]{geom_axes_lines()}}
}
\keyword{internal}
