% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_vert_line_args}
\alias{validate_vert_line_args}
\title{validate vertical line arguments}
\usage{
validate_vert_line_args(
  data,
  vline_arb = NULL,
  vline_arb_color = "red",
  vline_arb_label = NULL,
  vline_vars = NULL,
  vline_vars_colors = NULL,
  vline_vars_labels = NULL
)
}
\arguments{
\item{data}{data frame with variables which will be displayed in the plot.}

\item{vline_arb}{numeric value identifying intercept for arbitrary horizontal line.}

\item{vline_arb_color}{color for the arbitrary horizontal line that will appear on the plot.}

\item{vline_arb_label}{label for the arbitrary horizontal line that will appear on the legend.}

\item{vline_vars}{name(s) of variables `(ANR*)` or values `(*LOQ)` identifying intercept values.}

\item{vline_vars_colors}{color(s) for the horizontal lines defined by variables that will appear on the plot.}

\item{vline_vars_labels}{labels(s) for the horizontal lines defined by variables that will appear on the legend.}
}
\description{
validate vertical line arguments
}
