% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_axes_line.R
\name{vline_axis_label}
\alias{vline_axis_label}
\title{Add labels to the x axis equaling to the values of the lines}
\usage{
vline_axis_label(plot)
}
\arguments{
\item{plot}{('ggplot') a ggplot with either horizontal or vertical lines}
}
\value{
(\code{ggplot}) with the new axis breaks
}
\description{
Add labels to the x axis equaling to the values of the lines
}
\examples{

library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) +
geom_point() +
 goshawk:::geom_axes_lines(
   vline_arb = c(12, 25, 30),
   vline_arb_color = "red",
   vline_arb_label = "Vert Line"
 )
goshawk:::vline_axis_label(p)
}
