% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_horizontal_lines}
\alias{add_horizontal_lines}
\title{Add horizontal lines and their legend labels to a plot}
\usage{
add_horizontal_lines(
  plot,
  plot_data,
  agg_label = NULL,
  color_comb = NULL,
  new_hline_col = NULL,
  new_vline_col = NULL,
  hline_arb = NULL,
  hline_arb_color = "red",
  hline_arb_label = NULL,
  hline_vars = NULL,
  hline_vars_colors = NULL,
  hline_vars_labels = NULL,
  vline_arb = NULL,
  vline_arb_color = "green",
  vline_arb_label = NULL,
  vline_vars = NULL,
  vline_vars_colors = NULL,
  vline_vars_labels = NULL
)
}
\arguments{
\item{plot}{the ggplot2 plot object which the horizontal lines should be added to}

\item{plot_data}{data frame with variables which will be displayed in the plot.}

\item{agg_label}{label for the line denoting the Mean or Median.}

\item{color_comb}{character denoting the color of the Mean or Median line.}

\item{new_hline_col}{the name of the column to be added to plot_data that will hold the single value for
the arbitrary horizontal line}

\item{new_vline_col}{the name of the column to be added to plot_data that will hold the single value for
the arbitrary vertical line}

\item{hline_arb}{numeric value identifying intercept for arbitrary horizontal line.}

\item{hline_arb_color}{color for the arbitrary horizontal line that will appear on the plot.}

\item{hline_arb_label}{label for the arbitrary horizontal line that will appear on the legend.}

\item{hline_vars}{name(s) of variables `(ANR*)` or values `(*LOQ)` identifying intercept values.}

\item{hline_vars_colors}{color(s) for the horizontal lines defined by variables that will appear on the plot.}

\item{hline_vars_labels}{labels(s) for the horizontal lines defined by variables that will appear on the legend.}

\item{vline_arb}{numeric value identifying intercept for arbitrary vertical line.}

\item{vline_arb_color}{color for the arbitrary vertical line that will appear on the plot.}

\item{vline_arb_label}{label for the arbitrary vertical that will appear on the legend.}

\item{vline_vars}{name(s) of variables `(ANR*)` or values `(*LOQ)` identifying intercept values.}

\item{vline_vars_colors}{color(s) for the vertical lines defined by variables that will appear on the plot.}

\item{vline_vars_labels}{labels(s) for the vertical lines defined by variables that will appear on the legend.}
}
\description{
Add horizontal lines and their legend labels to a plot
}
