% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{caption_loqs_label}
\alias{caption_loqs_label}
\title{Helper for identifying the LLOQ and ULOQ values used in the visualization footnote.}
\usage{
caption_loqs_label(loqs_data)
}
\arguments{
\item{loqs_data}{dataset containing assay data with potential LOQ values}
}
\description{
Helper for identifying the LLOQ and ULOQ values used in the visualization footnote.
}
\details{
Biomarker Sciences would like to have LLOQ and ULOQ values available for
reference in the visualizations. This also aids in setting the data constraint
ranges when goshawk functions are run from teal.goshawk UI.
}
\examples{
library(scda)

ADLB <- synthetic_cdisc_data("latest")$adlb

# add footnote to identify LLOQ and ULOQ values pulled from data
caption_label <- goshawk:::caption_loqs_label(loqs_data = ADLB)

}
