% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_boxplot.R
\name{g_boxplot}
\alias{g_boxplot}
\title{Function to generate a boxplot
Output rendered by teal.goshawk module \code{g_boxplot} returns boxplot visualization}
\usage{
g_boxplot(data, biomarker, value_var, trt_group, loq_flag = "LOQFL",
  unit = NULL, timepoint = NULL, color_manual = NULL,
  shape_manual = NULL, box = TRUE, logscale = FALSE, ymax_scale = NULL,
  ymin_scale = NULL, dot_size = 2, alpha = 1, font_size = NULL,
  facet = NULL)
}
\arguments{
\item{data}{data frame with variables which will be displayed in the plot.}

\item{biomarker}{biomarker to visualize e.g. IGG.}

\item{value_var}{name of variable containing biomarker results displayed on
Y-axis e.g. AVAL.}

\item{trt_group}{name of variable representing treatment trt_group e.g. ARM.}

\item{loq_flag}{name of variable containing LOQ flag e.g. LBLOQFL.}

\item{unit}{biomarker unit label e.g. (U/L)}

\item{timepoint}{}

\item{color_manual}{vector of colour for trt_group}

\item{shape_manual}{vector of shapes (used with log_flag)}

\item{box}{add boxes to the plot (boolean)}

\item{logscale}{use a log scale for the Y axis (boolean)}

\item{ymax_scale}{maximum value for the Y axis}

\item{ymin_scale}{minimum value for the Y axis}

\item{alpha}{transparency for the points (0 = transparent, 1 = opaque)}

\item{font_size}{point size of tex to use.  NULL is use default size}

\item{facet}{variable to facet the plot by, or "None" if no faceting required.}
}
\value{
\code{ggplot} object
}
\description{
A box plot is a method for graphically depicting groups of numerical data
through their quartiles. Box plots may also have lines extending vertically
from the boxes (whiskers) indicating variability outside the upper and lower
quartiles, hence the term box-and-whisker. Outliers may be plotted as
individual points. Box plots are non-parametric: they display variation in
samples of a statistical population without making any assumptions of the
underlying statistical distribution. The spacings between the different parts
of the box indicate the degree of dispersion (spread) and skewness in the
data, and show outliers. In addition to the points themselves, they allow one
to visually estimate various L-estimators, notably the interquartile range,
midhinge, range, mid-range, and trimean.
}
\details{
provide additional information as needed. perhaps link to
  specification file.\url{http://rstudio.com}
}
\examples{

\dontrun{
# ALB is a VAD or ADAM format dataset, subsetted for the appropriate population.
# Select the parameter of interest to plot.
albplot <- ALB \%>\% 
  filter(PARAMCD == "IGA") 

# Plot boxplots for IGA by Treatment (ARM) faceted by AVISIT, to display a 
# boxplot for each visit.  
# Specific colors are given for each treatment group.
# Values that are flagged by LOQFL are given a different symbol.
g_boxplot(albplot
          , biomarker = "IGA"
          , value_var = "AVAL"
          , trt_group = "ARM"
          , loq_flag = 'LOQFL'
          , timepoint = "over time"
          , unit = "U/L"
          , color_manual = c('ARM 1' = "#1F78B4", 'ARM 2' = "#33A02C", 'ARM 3' = "#601010")
          , shape_manual = c('N' = 1, 'Y' = 2, 'NA' = NULL)
          , facet = "AVISIT"
          , box = TRUE
)

}


}
\author{
Balazs Toth

Jeff Tomlinson (tomlinsj) jeffrey.tomlinson@roche.com
}
