% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_density_distribution_plot.R
\name{g_density_distribution_plot}
\alias{g_density_distribution_plot}
\title{Function to generate a density distribution plot
Output rendered by teal.goshawk module \code{g_density_distribution_plot} returns
distribution overlay plot visualiztion}
\usage{
g_density_distribution_plot(label = "Density Distribution Plot", data = ALB,
  param_var = "PARAMCD", param = "CRP", xaxis_var = "AVAL",
  trt_group = "ARM", unit = "AVALU", color_manual = NULL, hline = NULL,
  rotate_xlab = FALSE, facet_var = "AVISITCD", font_size = 12,
  line_size = 2)
}
\arguments{
\item{label}{text string to used to identify plot.}

\item{data}{data frame with variables which will be displayed in the plot.
Note that the data are expected to be in vertical form with the
\code{PARAMCD} variable filtering to one observation per patient per visit.}

\item{param_var}{name of variable containing biomarker codes e.g. PARAMCD.}

\item{param}{biomarker to visualize e.g. IGG.}

\item{xaxis_var}{name of variable containing biomarker results displayed on X-axis e.g. BASE.}

\item{trt_group}{name of variable representing treatment group e.g. ARM.}

\item{unit}{name of variable containing biomarker unit e.g. AVALU.}

\item{color_manual}{vector of treatment colors. assigned values in app.R otherwise uses default colors.}

\item{hline}{y-axis value to position of horizontal line.}

\item{rotate_xlab}{45 degree rotation of x-axis values.}

\item{facet_var}{variable to use for facetting.}

\item{font_size}{control font size for title, x-axis, y-axis and legend font.}

\item{line_size}{plot line thickness.}
}
\description{
This function displays an overall density plot with treatment arms overlaid.
}
\details{
This function displays an overall density plot with treatment arms overlaid. link to specification file \url{http://rstudio.com}
}
\examples{

\dontrun{
# Example using analysis dataset for example ASL or ADSL,
# ALB points to biomarker data stored in a typical LB structure. for example ALB or ADLB.

# need a test data set created using random.cdisc.data.
# example call uses expects ALB structure 

param <- c('CRP') # FOR TESTING: woud come from teal.goshawk.tm_g_density_distribution_plot.R

plot1 <- g_density_distribution_plot(label = 'Density Distribution Plot',
           data = ALB,
           param_var = 'PARAMCD',
           param = param,
           xaxis_var = 'AVAL',
           trt_group = 'ARM',
           unit = 'AVALU',
           color_manual = color_manual,
           hline = NULL,
           rotate_xlab = FALSE,
           facet_var = 'AVISITCD',
           font_size = 10,
           line_size = .5)
plot1 

}
}
\author{
Nick Paszty (npaszty) paszty.nicholas@gene.com

Balazs Toth (tothb2)  toth.balazs@gene.com
}
