% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_density_distribution_plot.R
\name{g_density_distribution_plot}
\alias{g_density_distribution_plot}
\title{Function to create a density distribution plot.}
\usage{
g_density_distribution_plot(label = "Density Distribution Plot", data,
  param_var = "PARAMCD", param = "CRP", xaxis_var = "AVAL",
  trt_group = "ARM", unit = "AVALU", xmin = NA, xmax = NA,
  color_manual = NULL, color_comb = "#39ff14",
  facet_var = "AVISITCD", hline = NULL, facet_ncol = 2,
  rotate_xlab = FALSE, font_size = 12, line_size = 2)
}
\arguments{
\item{label}{text string used to identify plot.}

\item{data}{ADaM structured analysis laboratory data frame e.g. ALB.}

\item{param_var}{name of variable containing biomarker codes e.g. PARAMCD.}

\item{param}{biomarker to visualize e.g. IGG.}

\item{xaxis_var}{name of variable containing biomarker results displayed on X-axis e.g. AVAL.}

\item{trt_group}{name of variable representing treatment group e.g. ARM.}

\item{unit}{name of variable containing biomarker unit e.g. AVALU.}

\item{xmin}{x-axis lower zoom limit.}

\item{xmax}{x-axis upper zoom limit.}

\item{color_manual}{vector of colors applied to treatment values.}

\item{color_comb}{name or hex value for combined treatment color.}

\item{facet_var}{variable to use for facetting.}

\item{hline}{y-axis value to position a horizontal line.}

\item{facet_ncol}{number of facets per row.}

\item{rotate_xlab}{45 degree rotation of x-axis label values.}

\item{font_size}{font size control for title, x-axis label, y-axis label and legend.}

\item{line_size}{plot line thickness.}
}
\description{
Default plot displays overall density facetted by visit with treatment arms and combined 
treatment overlaid.
}
\examples{

\dontrun{
# Example using ADaM structure analysis dataset.

# original ARM value = dose value
arm_mapping <- list("A: Drug X" = "150mg QD", "B: Placebo" = "Placebo", 
"C: Combination" = "Combination")
color_manual <-  c("150mg QD" = "#000000", "Placebo" = "#3498DB", "Combination" = "#E74C3C")

library(dplyr)
library(ggplot2)
library(random.cdisc.data)
library(stringr)

ASL <- radsl(N = 20, seed = 1)
ALB <- radlb(ASL, visit_format = "WEEK", n_assessments = 7, seed = 2)
ALB <- ALB \%>\% 
mutate(AVISITCD = case_when(
AVISIT == "SCREENING" ~ "SCR",
AVISIT == "BASELINE" ~ "BL", grepl("WEEK", AVISIT) ~ paste("W",trimws(substr(AVISIT, start=6, 
stop=str_locate(AVISIT, "DAY")-1))),
TRUE ~ as.character(NA))) \%>\%
mutate(AVISITCDN = case_when(AVISITCD == "SCR" ~ -2,
AVISITCD == "BL" ~ 0, grepl("W", AVISITCD) ~ as.numeric(gsub("\\\\D+", "", AVISITCD)), 
TRUE ~ as.numeric(NA))) \%>\%
# use ARMCD values to order treatment in visualization legend
mutate(TRTORD = ifelse(grepl("C", ARMCD), 1,
ifelse(grepl("B", ARMCD), 2,
ifelse(grepl("A", ARMCD), 3, NA)))) \%>\%
mutate(ARM = as.character(arm_mapping[match(ARM, names(arm_mapping))])) \%>\%
mutate(ARM = factor(ARM) \%>\% reorder(TRTORD))

param <- c('CRP')

plot1 <- g_density_distribution_plot(label = 'Density Distribution Plot',
           data = ALB,
           param_var = 'PARAMCD',
           param = param,
           xaxis_var = 'AVAL',
           trt_group = 'ARM',
           xmin = 0,
           xmax = 200,
           unit = 'AVALU',
           color_manual = color_manual,
           color_comb = "#39ff14",
           facet_var = 'AVISITCD',
           hline = NULL,
           facet_ncol = 2,
           rotate_xlab = FALSE,
           font_size = 10,
           line_size = .5)
plot1 

}

}
\author{
Nick Paszty (npaszty) paszty.nicholas@gene.com

Balazs Toth (tothb2)  toth.balazs@gene.com
}
