% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_density_distribution_plot.R
\name{g_density_distribution_plot}
\alias{g_density_distribution_plot}
\title{Function to create a density distribution plot.}
\usage{
g_density_distribution_plot(label = "Density Distribution Plot", data,
  param_var = "PARAMCD", param = "CRP", xaxis_var = "AVAL",
  trt_group = "ARM", unit = "AVALU", xmin = NA, xmax = NA,
  color_manual = NULL, facet_var = "AVISITCD", hline = NULL,
  facet_ncol = 2, rotate_xlab = FALSE, font_size = 12, line_size = 2)
}
\arguments{
\item{label}{text string used to identify plot.}

\item{data}{ADaM structured analysis laboratory data frame e.g. ALB.}

\item{param_var}{name of variable containing biomarker codes e.g. PARAMCD.}

\item{param}{biomarker to visualize e.g. IGG.}

\item{xaxis_var}{name of variable containing biomarker results displayed on X-axis e.g. AVAL.}

\item{trt_group}{name of variable representing treatment group e.g. ARM.}

\item{unit}{name of variable containing biomarker unit e.g. AVALU.}

\item{xmin}{x-axis lower zoom limit.}

\item{xmax}{x-axis upper zoom limit.}

\item{color_manual}{vector of colors applied to treatment values.}

\item{facet_var}{variable to use for facetting.}

\item{hline}{y-axis value to position a horizontal line.}

\item{facet_ncol}{number of facets per row.}

\item{rotate_xlab}{45 degree rotation of x-axis label values.}

\item{font_size}{font size control for title, x-axis label, y-axis label and legend.}

\item{line_size}{plot line thickness.}
}
\description{
Default plot displays overall density facetted by visit with treatment arms and combined treatment overlaid.
}
\examples{

\dontrun{
# Example using ADaM structure analysis dataset.
# ALB refers to biomarker data stored in expected laboratory structure.

param <- c('CRP')
color_manual <-  c('Placebo' = "#000000", '150mg QD' = "#3498DB", '200mg BID' = "#E74C3C")
ALBsub <- ALB \%>\% subset(AVISITCDN ==24)

plot1 <- g_density_distribution_plot(label = 'Density Distribution Plot',
           data = ALBsub,
           param_var = 'PARAMCD',
           param = param,
           xaxis_var = 'AVAL',
           trt_group = 'ARM',
           xmin = 0,
           xmax = 200,
           unit = 'AVALU',
           color_manual = color_manual,
           facet_var = 'AVISITCD',
           hline = NULL,
           facet_ncol = 2,
           rotate_xlab = FALSE,
           font_size = 10,
           line_size = .5)
plot1 

}

}
\author{
Nick Paszty (npaszty) paszty.nicholas@gene.com

Balazs Toth (tothb2)  toth.balazs@gene.com
}
