% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_lineplot.R
\name{g_lineplot}
\alias{g_lineplot}
\title{Function to generate a line plot
output rendered by teal.goshawk module}
\usage{
g_lineplot(label = "Line Plot", data, biomarker_var = "PARAMCD", biomarker,
  value_var = "AVAL", trt_group, time, color_manual, median = FALSE,
  hline = NULL, rotate_xlab = FALSE)
}
\arguments{
\item{label}{text string to be displayed as plot label.}

\item{data}{data frame with variables to be summarized and generate statistics which will display in the plot.}

\item{biomarker_var}{name of variable containing biomarker names.}

\item{value_var}{name of variable containing biomarker results.}

\item{trt_group}{name of variable representing treatment group.}

\item{time}{name of vairable containing visit names.}

\item{color_manual}{vector of colors.}

\item{median}{boolean whether to display median results.}

\item{hline}{numeric value represnting intercept of horizontal line.}

\item{biomaker}{biomarker name to be analyzed.}

\item{roate_xlab}{boolean whether to rotate x-axis labels.}
}
\value{
\code{ggplot} object
}
\description{
\code{g_lineplot} returns a ggplot object
}
\details{
provide additional information as needed. perhaps link to specification file.\url{http://rstudio.com}
}
\examples{


}
\author{
Balazs Toth

Wenyi Liu (luiw2) wenyi.liu@roche.com
}
