% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_scatterplot.R
\name{g_scatterplot}
\alias{g_scatterplot}
\title{Function to generate a scatter plot
Output rendered by teal.goshawk module \code{g_scatterplot} returns scatter plot visualiztion}
\usage{
g_scatterplot(data, biomarker, biomarker.bl, trt_group, time, loq_flag, unti,
  timepoint, color_manual, shape_manual, logscale = FALSE, facet = NULL,
  pct = FALSE, reg.line = FALSE)
}
\arguments{
\item{add}{as many param lines as there are params that this function expects}

\item{add}{as many param lines as there are params that this function expects}
}
\value{
\code{scatterplot} object
}
\description{
Function to generate a scatter plot
Output rendered by teal.goshawk module \code{g_scatterplot} returns scatter plot visualiztion
}
\details{
provide additional information as needed. link to specification file \url{http://rstudio.com}
}
\examples{
# Create dataset
biomarker <- runif(100,3,100)
biomarker.bl <- NULL
 for (i in 0:24) {
   biomarker.bl <- c ( biomarker.bl, rep(biomarker[i * 4 + 1], 4))
 }
arm <- c(rep('trt',50),rep('pbo',50))
visit <- c(rep(c(0,4,8,12),25))
loq_flag <- c(rep(c(rep('YES',10),rep('NO',10)),5))
example <- data.frame(biomarker = biomarker, biomarker.bl = biomarker.bl, arm = arm, loq_flag = loq_flag, visit = visit)
colnames(example) <- c('IGG', 'IGG.bl', 'arm','loq_flag','visit')
unit <- 'g/L'
timepoint <- 'screening'
color_manual <- c('pbo' = "#1F78B4", 'trt' = "#33A02C")
shape_manual <- c('NO' = 1, 'YES' = 2, 'NA' = 0)
# Call function simple
g_scatterplot (data = example,
                  biomarker = 'IGG',
                  biomarker.bl = 'IGG.bl',
                  trt_group = 'arm',
                  time = 'visit',
                  loq_flag = 'loq_flag',
                  unit,
                  timepoint,
                  color_manual,
                  shape_manual)

# Call function facet
g_scatterplot (data = example,
                  biomarker = 'IGG',
                  biomarker.bl = 'IGG.bl',
                  trt_group = 'arm',
                  time = 'visit',
                  loq_flag = 'loq_flag',
                  unit,
                  timepoint,
                  color_manual,
                  shape_manual,
                  logscale = TRUE,
                  facet = 'arm',
                  pct = FALSE,
                  reg.line = FALSE)

# Call function facet and logscale
g_scatterplot (data = example,
                  biomarker = 'IGG',
                  biomarker.bl = 'IGG.bl',
                  trt_group = 'arm',
                  time = 'visit',
                  loq_flag = 'loq_flag',
                  unit,
                  timepoint,
                  color_manual,
                  shape_manual,
                  logscale = TRUE,
                  facet = 'arm',
                  pct = FALSE,
                  reg.line = FALSE)

# Call function facet and pct
g_scatterplot (data = example,
                  biomarker = 'IGG',
                  biomarker.bl = 'IGG.bl',
                  trt_group = 'arm',
                  time = 'visit',
                  loq_flag = 'loq_flag',
                  unit,
                  timepoint,
                  color_manual,
                  shape_manual,
                  logscale = FALSE,
                  facet = 'arm',
                  pct = TRUE,
                  reg.line = FALSE)

# Call function regression line
g_scatterplot (data = example,
                  biomarker = 'IGG',
                  biomarker.bl = 'IGG.bl',
                  trt_group = 'arm',
                  time = 'visit',
                  loq_flag = 'loq_flag',
                  unit,
                  timepoint,
                  color_manual,
                  shape_manual,
                  logscale = FALSE,
                  facet = 'arm',
                  pct = FALSE,
                  reg.line = TRUE)

# Call function regression line
g_scatterplot (data = example,
                  biomarker = 'IGG',
                  biomarker.bl = 'IGG.bl',
                  trt_group = 'arm',
                  time = 'visit',
                  loq_flag = 'loq_flag',
                  unit,
                  timepoint,
                  color_manual,
                  shape_manual,
                  logscale = FALSE,
                  facet = 'arm',
                  pct = TRUE,
                  reg.line = TRUE)

# Call function regression line
g_scatterplot (data = example,
                  biomarker = 'IGG',
                  biomarker.bl = 'IGG.bl',
                  trt_group = 'arm',
                  time = 'visit',
                  loq_flag = 'loq_flag',
                  unit,
                  timepoint,
                  color_manual,
                  shape_manual,
                  logscale = FALSE,
                  facet = 'arm',
                  pct = FALSE,
                  reg.line = TRUE)


}
\author{
Balazs Toth

Nick Paszty
}
