% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_scatterplot.R
\name{g_scatterplot}
\alias{g_scatterplot}
\title{Function to create a scatter plot.}
\usage{
g_scatterplot(...)
}
\arguments{
\item{...}{function is deprecated.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{g_scatterplot()} is deprecated.
Please use \code{\link[=g_correlationplot]{g_correlationplot()}} instead.
}
