% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_scatterplot.R
\name{g_scatterplot}
\alias{g_scatterplot}
\title{Function to generate a scatter plot
Output rendered by teal.goshawk module \code{g_scatterplot} returns scatter plot visualiztion}
\usage{
g_scatterplot(label = "Scatter Plot", data = ALB, param_var = "PARAMCD",
  param = "CRP", xaxis_var = "BASE", yaxis_var = "AVAL",
  trt_group = "ARM", visit = "AVISITCD", loq_flag_var = "LOQFL",
  unit = "AVALU", xmin_scale = 0, xmax_scale = 200, ymin_scale = 0,
  ymax_scale = 200, color_manual = NULL, shape_manual = c(N = 1, Y = 2,
  `NA` = 0), facet = FALSE, facet_var = "ARM", reg_line = FALSE,
  rotate_xlab = FALSE, hline = NULL, font_size = 12, dot_size = NULL,
  reg_text_size = 3)
}
\arguments{
\item{label}{text string to used to identify plot.}

\item{data}{data frame with variables which will be displayed in the plot.
Note that the data are expected to be in vertical form with the
\code{PARAMCD} variable filtering to one observation per patient per visit.}

\item{param_var}{name of variable containing biomarker codes e.g. PARAMCD.}

\item{param}{biomarker to visualize e.g. IGG.}

\item{xaxis_var}{name of variable containing biomarker results displayed on X-axis e.g. BASE.}

\item{yaxis_var}{name of variable containing biomarker results displayed on Y-axise.g. AVAL.}

\item{trt_group}{name of variable representing treatment group e.g. ARM.}

\item{visit}{name of variable containing nominal visits e.g. AVISITCD.}

\item{loq_flag_var}{name of variable containing LOQ flag e.g. LBLOQFL.}

\item{unit}{name of variable containing biomarker unit e.g. AVALU.}

\item{xmin_scale}{minimum value of xaxis_var variable for selected biomarker.}

\item{xmax_scale}{maximum value of xaxis_var variable for selected biomarker.}

\item{ymin_scale}{minimum value of yaxis_var variable for selected biomarker.}

\item{ymax_scale}{maximum value of yaxis_var variable for selected biomarker.}

\item{color_manual}{vector of colors. Currently plot uses default colors and this control is set to NULL.}

\item{shape_manual}{vector of shapes to display LOQ values vs. non-LOQ values.}

\item{facet}{set layout to use facetting.}

\item{facet_var}{variable to use for facetting.}

\item{reg_line}{include regression line and annotations for slope and coefficient in visualization. Use with facet TRUE.}

\item{rotate_xlab}{45 degree rotation of x-axis values.}

\item{hline}{y-axis value to position of horizontal line.}

\item{font_size}{control font size for title, x-axis, y-axis and legend font.}

\item{dot_size}{scatter dot size.}

\item{reg_text_size}{regression line annotation font size.}
}
\description{
A scatter plot is a type of plot using Cartesian coordinates to display values for typically two variables or
for one variable at different timepoints for a set of data.
If the points are color-coded, one additional variable can be displayed.
The data are displayed as a collection of points, each having the value of one variable or timepoint determining
the position on the horizontal axis and the value of the other variable or timepoint determining the position 
on the vertical axis.
}
\details{
This function displays a scatter plot. link to specification file \url{http://rstudio.com}
}
\examples{

\dontrun{
# Example using analysis dataset for example ASL or ADSL,
# ALB points to biomarker data stored in a typical LB structure. for example ALB or ADLB.

# need a test data set created using random.cdisc.data.
# example call uses expects ALB structure 

param <- c('CRP') # FOR TESTING: woud come from teal.goshawk.tm_g_scatterplot.R

plot1 <- g_scatterplot(label = 'Scatter Plot',
           data = ALB,
           param_var = 'PARAMCD', 
           param = param,
           xaxis_var = 'BASE',
           yaxis_var = 'AVAL',
           trt_group = 'ARM',
           visit = 'AVISITCD',
           loq_flag_var = 'LOQFL',
           unit = 'AVALU',
           xmin_scale = 0,
           xmax_scale = .5,
           ymin_scale = 0,
           ymax_scale = .5,
           color_manual = NULL,
           shape_manual = c('N' = 1, 'Y' = 2, 'NA' = 0),
           hline = NULL,
           rotate_xlab = FALSE,
           facet = FALSE,
           facet_var = "ARM",
           reg_line = FALSE,
           font_size = 14,
           dot_size = 2,
           reg_text_size = 3)
plot1 

}
}
\author{
Nick Paszty (npaszty) paszty.nicholas@gene.com

Balazs Toth (tothb2)  toth.balazs@gene.com
}
