% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_scatterplot.R
\name{g_scatterplot}
\alias{g_scatterplot}
\title{Function to create a scatter plot.}
\usage{
g_scatterplot(
  label = "Scatter Plot",
  data,
  param_var = "PARAMCD",
  param = "CRP",
  xaxis_var = "BASE",
  yaxis_var = "AVAL",
  trt_group = "ARM",
  visit = "AVISITCD",
  loq_flag_var = "LOQFL",
  unit = "AVALU",
  xmin = NA,
  xmax = NA,
  ymin = NA,
  ymax = NA,
  color_manual = NULL,
  shape_manual = NULL,
  facet_ncol = 2,
  facet = FALSE,
  facet_var = "ARM",
  reg_line = FALSE,
  hline = NULL,
  vline = NULL,
  rotate_xlab = FALSE,
  font_size = 12,
  dot_size = NULL,
  reg_text_size = 3
)
}
\arguments{
\item{label}{text string to used to identify plot.}

\item{data}{ADaM structured analysis laboratory data frame e.g. ALB.}

\item{param_var}{name of variable containing biomarker codes e.g. PARAMCD.}

\item{param}{biomarker to visualize e.g. IGG.}

\item{xaxis_var}{name of variable containing biomarker results displayed on X-axis e.g. BASE.}

\item{yaxis_var}{name of variable containing biomarker results displayed on Y-axise.g. AVAL.}

\item{trt_group}{name of variable representing treatment group e.g. ARM.}

\item{visit}{name of variable containing nominal visits e.g. AVISITCD.}

\item{loq_flag_var}{name of variable containing LOQ flag e.g. LOQFL.}

\item{unit}{name of variable containing biomarker unit e.g. AVALU.}

\item{xmin}{x-axis lower zoom limit.}

\item{xmax}{x-axis upper zoom limit.}

\item{ymin}{y-axis lower zoom limit.}

\item{ymax}{y-axis upper zoom limit.}

\item{color_manual}{vector of colors applied to treatment values.}

\item{shape_manual}{vector of symbols applied to LOQ values.}

\item{facet_ncol}{number of facets per row.}

\item{facet}{set layout to use treatment facetting.}

\item{facet_var}{variable to use for treatment facetting.}

\item{reg_line}{include regression line and annotations for slope and coefficient in
visualization. Use with facet = TRUE.}

\item{hline}{y-axis value to position a horizontal line.}

\item{vline}{x-axis value to position a vertical line.}

\item{rotate_xlab}{45 degree rotation of x-axis label values.}

\item{font_size}{font size control for title, x-axis label, y-axis label and legend.}

\item{dot_size}{plot dot size.}

\item{reg_text_size}{font size control for regression line annotations.}
}
\description{
Default plot displays scatter facetted by visit with color attributed treatment arms and symbol
attributed LOQ values.
}
\details{
Regression uses deming model.
}
\examples{
# Example using ADaM structure analysis dataset.

library(scda)
library(stringr)

# original ARM value = dose value
arm_mapping <- list("A: Drug X" = "150mg QD", "B: Placebo" = "Placebo",
"C: Combination" = "Combination")
color_manual <-  c("150mg QD" = "#000000", "Placebo" = "#3498DB", "Combination" = "#E74C3C")
# assign LOQ flag symbols: circles for "N" and triangles for "Y", squares for "NA"
shape_manual <-  c("N" = 1, "Y" = 2, "NA" = 0)

ASL <- synthetic_cdisc_data("latest")$adsl
ALB <- synthetic_cdisc_data("latest")$adlb
var_labels <- lapply(ALB, function(x) attributes(x)$label)
ALB <- ALB \%>\%
  mutate(AVISITCD = case_when(
    AVISIT == "SCREENING" ~ "SCR",
    AVISIT == "BASELINE" ~ "BL",
    grepl("WEEK", AVISIT) ~
      paste(
        "W",
        trimws(
          substr(
            AVISIT,
            start = 6,
            stop = str_locate(AVISIT, "DAY") - 1
          )
        )
      ),
    TRUE ~ NA_character_)) \%>\%
  mutate(AVISITCDN = case_when(
    AVISITCD == "SCR" ~ -2,
    AVISITCD == "BL" ~ 0,
    grepl("W", AVISITCD) ~ as.numeric(gsub("\\\\D+", "", AVISITCD)),
    TRUE ~ NA_real_)) \%>\%
  # use ARMCD values to order treatment in visualization legend
  mutate(TRTORD = ifelse(grepl("C", ARMCD), 1,
    ifelse(grepl("B", ARMCD), 2,
      ifelse(grepl("A", ARMCD), 3, NA)))) \%>\%
  mutate(ARM = as.character(arm_mapping[match(ARM, names(arm_mapping))])) \%>\%
  mutate(ARM = factor(ARM) \%>\%
  reorder(TRTORD))
 attr(ALB[["ARM"]], "label") <- var_labels[["ARM"]]

g_scatterplot(label = "Scatter Plot",
           data = ALB,
           param_var = "PARAMCD",
           param = c("ALT"),
           xaxis_var = "BASE",
           yaxis_var = "AVAL",
           trt_group = "ARM",
           visit = "AVISITCD",
           loq_flag_var = "LOQFL",
           unit = "AVALU",
           xmin = 0,
           xmax = 200,
           ymin = 0,
           ymax = 200,
           color_manual = color_manual,
           shape_manual = shape_manual,
           facet_ncol = 2,
           facet = TRUE,
           facet_var = "ARM",
           reg_line = TRUE,
           hline = NULL,
           vline = .5,
           rotate_xlab = TRUE,
           font_size = 14,
           dot_size = 2,
           reg_text_size = 3)
}
\author{
Nick Paszty (npaszty) paszty.nicholas@gene.com

Balazs Toth (tothb2)  toth.balazs@gene.com
}
