% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_scatterplot.R
\name{g_scatterplot}
\alias{g_scatterplot}
\title{Function to create a scatter plot.}
\usage{
g_scatterplot(label = "Scatter Plot", data = ALB, param_var = "PARAMCD",
  param = "CRP", xaxis_var = "BASE", yaxis_var = "AVAL",
  trt_group = "ARM", visit = "AVISITCD", loq_flag_var = "LOQFL",
  unit = "AVALU", color_manual = NULL, shape_manual = NULL,
  facet = FALSE, facet_var = "ARM", reg_line = FALSE, hline = NULL,
  rotate_xlab = FALSE, font_size = 12, dot_size = NULL,
  reg_text_size = 3)
}
\arguments{
\item{label}{text string to used to identify plot.}

\item{data}{ADaM structured analysis laboratory data frame e.g. ALB.}

\item{param_var}{name of variable containing biomarker codes e.g. PARAMCD.}

\item{param}{biomarker to visualize e.g. IGG.}

\item{xaxis_var}{name of variable containing biomarker results displayed on X-axis e.g. BASE.}

\item{yaxis_var}{name of variable containing biomarker results displayed on Y-axise.g. AVAL.}

\item{trt_group}{name of variable representing treatment group e.g. ARM.}

\item{visit}{name of variable containing nominal visits e.g. AVISITCD.}

\item{loq_flag_var}{name of variable containing LOQ flag e.g. LBLOQFL.}

\item{unit}{name of variable containing biomarker unit e.g. AVALU.}

\item{color_manual}{vector of colors applied to treatment values.}

\item{shape_manual}{vector of symbols applied to LOQ values.}

\item{facet}{set layout to use treatment facetting.}

\item{facet_var}{variable to use for facetting beyond visit.}

\item{reg_line}{include regression line and annotations for slope and coefficient in visualization. Use with facet = TRUE.}

\item{hline}{y-axis value to position a horizontal line.}

\item{rotate_xlab}{45 degree rotation of x-axis label values.}

\item{font_size}{font size control for title, x-axis label, y-axis label and legend.}

\item{dot_size}{plot dot size.}

\item{reg_text_size}{font size control for regression line annotations.}
}
\description{
Default plot displays scatter facetted by visit with color attributed treatment arms and symbol attributed LOQ values.
}
\details{
Regression uses deming model.
}
\examples{

\dontrun{
# Example using ADaM structure analysis dataset.
# ALB refers to biomarker data stored in expected laboratory structure.

param <- c('ADIGG') # FOR TESTING: woud come from teal.goshawk.tm_g_scatterplot.R

plot1 <- g_scatterplot(label = 'Scatter Plot',
           data = ALB,
           param_var = 'PARAMCD', 
           param = param,
           xaxis_var = 'BASE',
           yaxis_var = 'AVAL',
           trt_group = 'ARM',
           visit = 'AVISITCD',
           loq_flag_var = 'LOQFL',
           unit = 'AVALU',
           color_manual = color_manual,
           shape_manual = shape_manual,
           facet = TRUE,
           facet_var = "ARM",
           reg_line = TRUE,
           hline = NULL,
           rotate_xlab = FALSE,
           font_size = 14,
           dot_size = 2,
           reg_text_size = 3)
plot1 

}

}
\author{
Nick Paszty (npaszty) paszty.nicholas@gene.com

Balazs Toth (tothb2)  toth.balazs@gene.com
}
