% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_spaghettiplot.R
\name{g_spaghettiplot}
\alias{g_spaghettiplot}
\title{Function to create a spaghetti plot.}
\usage{
g_spaghettiplot(data, subj_id = "USUBJID", biomarker_var = "PARAMCD",
  biomarker_var_label = "PARAM", biomarker, value_var = "AVAL",
  unit_var = "AVALU", trt_group, trt_group_level = NULL, time,
  time_level = NULL, color_manual = NULL, ylim = NULL, alpha = 1,
  facet_ncol = 2, hline = NULL, xtick = waiver(), xlabel = xtick,
  rotate_xlab = FALSE, font_size = 12, group_stats = "NONE")
}
\arguments{
\item{data}{data frame with variables to be summarized and generate statistics which will display in the plot.}

\item{subj_id}{unique subject id variable name.}

\item{biomarker_var}{name of variable containing biomarker names.}

\item{biomarker_var_label}{name of variable containing biomarker labels.}

\item{biomarker}{biomarker name to be analyzed.}

\item{value_var}{name of variable containing biomarker results.}

\item{unit_var}{name of variable containing biomarker units.}

\item{trt_group}{name of variable representing treatment group.}

\item{trt_group_level}{vector that can be used to define the factor level of trt_group.}

\item{time}{name of vairable containing visit names.}

\item{time_level}{vector that can be used to define the factor level of time. Only use it when x-axis variable is character or factor.}

\item{color_manual}{vector of colors.}

\item{ylim}{numeric vector to define y-axis range.}

\item{alpha}{subject line transparency (0 = transparent, 1 = opaque)}

\item{facet_ncol}{number of facets per row.}

\item{hline}{numeric value represnting intercept of horizontal line.}

\item{xtick}{numeric vector to define the tick values of x-axis when x variable is numeric. Default value is waiver().}

\item{xlabel}{vector with same length of xtick to define the label of x-axis tick values. Default value is waiver().}

\item{rotate_xlab}{boolean whether to rotate x-axis labels.}

\item{font_size}{control font size for title, x-axis, y-axis and legend font.}

\item{group_stats}{control group mean or median overlay.}
}
\value{
\code{ggplot} object
}
\description{
This function is rendered by teal.goshawk module
}
\examples{

\dontrun{
# EXAMPLE:

library(dplyr)

ANL <- expand.grid(
  USUBJID = paste0("p-",1:100),
  VISITN = c(1, 4:10),
  ARM = c("ARM A", "ARM B", "ARM C"),
  PARAMCD = c("CRP", "IGG", "IGM"),
  PARAM = c("C-reactive protein", "Immunoglobulin G", "Immunoglobulin M")
)
ANL$VISIT <- paste0("visit ", ANL$VISITN)
ANL$AVAL <- rnorm(nrow(ANL))
ANL$CHG <- rnorm(nrow(ANL), 2, 2)
ANL$CHG[ANL$VISIT == "visit 1"] <- NA
ANL$PCHG <- ANL$CHG/ANL$AVAL*100
ANL$AVALU <- 'mg'

ANL$ARM <- factor(ANL$ARM)
ANL$VISIT <- factor(ANL$VISIT)

g_spaghettiplot(data = ANL,
                subj_id = 'USUBJID',
                biomarker_var = 'PARAMCD',
                biomarker = 'CRP',
                value_var = 'AVAL',
                trt_group = 'ARM',
                time = 'VISIT',
                color_manual = c("ARM A" = "#000000", "ARM B" = "#3498DB", "ARM C" = "#E74C3C"),
                alpha = .02,
                hline = NULL,
                rotate_xlab = FALSE,
                group_stats = "median"
                )

}

}
\author{
Wenyi Liu (wenyi.liu@roche.com)
}
