% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_spaghettiplot.R
\name{g_spaghettiplot}
\alias{g_spaghettiplot}
\title{Function to create a spaghetti plot.}
\usage{
g_spaghettiplot(
  data,
  subj_id = "USUBJID",
  biomarker_var = "PARAMCD",
  biomarker_var_label = "PARAM",
  biomarker,
  value_var = "AVAL",
  unit_var = "AVALU",
  trt_group,
  trt_group_level = NULL,
  time,
  time_level = NULL,
  color_manual = NULL,
  color_comb = "#39ff14",
  ylim = NULL,
  alpha = 1,
  facet_ncol = 2,
  xtick = waiver(),
  xlabel = xtick,
  rotate_xlab = FALSE,
  font_size = 12,
  group_stats = "NONE",
  hline_arb = NULL,
  hline_arb_color = "red",
  hline_arb_label = NULL,
  hline_vars = NULL,
  hline_vars_colors = NULL,
  hline_vars_labels = NULL
)
}
\arguments{
\item{data}{data frame with variables to be summarized and generate statistics which will display
in the plot.}

\item{subj_id}{unique subject id variable name.}

\item{biomarker_var}{name of variable containing biomarker names.}

\item{biomarker_var_label}{name of variable containing biomarker labels.}

\item{biomarker}{biomarker name to be analyzed.}

\item{value_var}{name of variable containing biomarker results.}

\item{unit_var}{name of variable containing biomarker units.}

\item{trt_group}{name of variable representing treatment group.}

\item{trt_group_level}{vector that can be used to define the factor level of trt_group.}

\item{time}{name of variable containing visit names.}

\item{time_level}{vector that can be used to define the factor level of time. Only use it when
x-axis variable is character or factor.}

\item{color_manual}{vector of colors.}

\item{color_comb}{name or hex value for combined treatment color.}

\item{ylim}{numeric vector to define y-axis range.}

\item{alpha}{subject line transparency (0 = transparent, 1 = opaque)}

\item{facet_ncol}{number of facets per row.}

\item{xtick}{a vector to define the tick values of time in x-axis.
Default value is waiver().}

\item{xlabel}{vector with same length of xtick to define the label of x-axis tick values. Default
value is waiver().}

\item{rotate_xlab}{boolean whether to rotate x-axis labels.}

\item{font_size}{control font size for title, x-axis, y-axis and legend font.}

\item{group_stats}{control group mean or median overlay.}

\item{hline_arb}{numeric value identifying intercept for arbitrary horizontal line.}

\item{hline_arb_color}{color for hline_arb that will appear on the plot.}

\item{hline_arb_label}{label for hline_arb that will appear on the legend.}

\item{hline_vars}{name(s) of variables (ANR*) or values (*LOQ) identifying interecpt values.}

\item{hline_vars_colors}{color(s) for the lines of hline_arb that will appear on the plot.}

\item{hline_vars_labels}{labels(s) for hline_arb that will appear on the legend.}
}
\value{
\code{ggplot} object
}
\description{
This function is rendered by teal.goshawk module
}
\examples{

# Example using ADaM structure analysis dataset.

library(scda)
library(stringr)

# original ARM value = dose value
arm_mapping <- list("A: Drug X" = "150mg QD", "B: Placebo" = "Placebo",
"C: Combination" = "Combination")
color_manual <-  c("150mg QD" = "#000000", "Placebo" = "#3498DB", "Combination" = "#E74C3C")

ASL <- synthetic_cdisc_data("latest")$adsl
ALB <- synthetic_cdisc_data("latest")$adlb
var_labels <- lapply(ALB, function(x) attributes(x)$label)
ALB <- ALB \%>\%
  mutate(AVISITCD = case_when(
    AVISIT == "SCREENING" ~ "SCR",
    AVISIT == "BASELINE" ~ "BL",
    grepl("WEEK", AVISIT) ~
      paste(
        "W",
        trimws(
          substr(
            AVISIT,
            start = 6,
            stop = str_locate(AVISIT, "DAY") - 1
          )
        )
      ),
    TRUE ~ NA_character_)) \%>\%
  mutate(AVISITCDN = case_when(
    AVISITCD == "SCR" ~ -2,
    AVISITCD == "BL" ~ 0,
    grepl("W", AVISITCD) ~ as.numeric(gsub("\\\\D+", "", AVISITCD)),
    TRUE ~ NA_real_)) \%>\%
  # use ARMCD values to order treatment in visualization legend
  mutate(TRTORD = ifelse(grepl("C", ARMCD), 1,
    ifelse(grepl("B", ARMCD), 2,
      ifelse(grepl("A", ARMCD), 3, NA)))) \%>\%
  mutate(ARM = as.character(arm_mapping[match(ARM, names(arm_mapping))])) \%>\%
  mutate(ARM = factor(ARM) \%>\%
  reorder(TRTORD)) \%>\%
  mutate(ANRLO = 30, ANRHI = 75) \%>\%
  rowwise() \%>\%
  group_by(PARAMCD) \%>\%
  mutate(LBSTRESC = ifelse(USUBJID \%in\% sample(USUBJID, 1, replace = TRUE),
  paste("<", round(runif(1, min = 25, max = 30))), LBSTRESC)) \%>\%
  mutate(LBSTRESC = ifelse(USUBJID \%in\% sample(USUBJID, 1, replace = TRUE),
  paste( ">", round(runif(1, min = 70, max = 75))), LBSTRESC)) \%>\%
  ungroup
  attr(ALB[["ARM"]], "label") <- var_labels[["ARM"]]
  attr(ALB[["ANRLO"]], "label") <- "Analysis Normal Range Lower Limit"
  attr(ALB[["ANRHI"]], "label") <- "Analysis Normal Range Upper Limit"

  # add LLOQ and ULOQ variables
  ALB_LOQS <- goshawk:::h_identify_loq_values(ALB)
  ALB <- left_join(ALB, ALB_LOQS, by = "PARAM")

g_spaghettiplot(data = ALB,
                subj_id = "USUBJID",
                biomarker_var = "PARAMCD",
                biomarker = "CRP",
                value_var = "AVAL",
                trt_group = "ARM",
                time = "AVISITCD",
                color_manual = color_manual,
                color_comb = "#39ff14",
                alpha = .02,
                xtick = c("BL", "W 1", "W 4"),
                xlabel = c("Baseline", "Week 1", "Week 4"),
                rotate_xlab = FALSE,
                group_stats = "median",
                hline_arb = NULL,
                hline_arb_color = NULL,
                hline_arb_label = NULL,
                hline_vars = c("ANRHI", "ANRLO"),
                hline_vars_colors = c("pink", "brown"),
                hline_vars_labels = NULL,
                )

g_spaghettiplot(data = ALB,
                subj_id = "USUBJID",
                biomarker_var = "PARAMCD",
                biomarker = "CRP",
                value_var = "AVAL",
                trt_group = "ARM",
                time = "AVISITCD",
                color_manual = color_manual,
                color_comb = "#39ff14",
                alpha = .02,
                xtick = c("BL", "W 1", "W 4"),
                xlabel = c("Baseline", "Week 1", "Week 4"),
                rotate_xlab = FALSE,
                group_stats = "median",
                hline_arb = NULL,
                hline_arb_color = NULL,
                hline_arb_label = NULL,
                hline_vars = c("ANRHI", "ANRLO", "LLOQ_VALUE_N", "ULOQ_VALUE_N"),
                hline_vars_colors = c("pink", "brown", "purple", "gray"),
                hline_vars_labels = NULL,
                )

g_spaghettiplot(data = ALB,
                subj_id = "USUBJID",
                biomarker_var = "PARAMCD",
                biomarker = "CRP",
                value_var = "AVAL",
                trt_group = "ARM",
                time = "AVISITCDN",
                color_manual = color_manual,
                color_comb = "#39ff14",
                alpha = .02,
                xtick = c(0, 1, 4),
                xlabel = c("Baseline", "Week 1", "Week 4"),
                rotate_xlab = FALSE,
                group_stats = "median",
                hline_arb = NULL,
                hline_arb_color = NULL,
                hline_arb_label = NULL,
                hline_vars = c("ANRLO", "ANRHI"),
                hline_vars_colors = NULL,
                hline_vars_labels = NULL,
                )

}
\author{
Wenyi Liu (wenyi.liu@roche.com)
}
