% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goshawk.R
\docType{package}
\name{goshawk-package}
\alias{goshawk}
\alias{goshawk-package}
\title{The goshawk package provides longitudinal visualizations of lab/biomarker data}
\description{
The goshawk package provides longitudinal visualizations of lab/biomarker data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightsengineering.github.io/goshawk/}
  \item \url{https://github.com/insightsengineering/goshawk/}
  \item Report bugs at \url{https://github.com/insightsengineering/goshawk/issues}
}

}
\author{
\strong{Maintainer}: Nick Paszty \email{nick.paszty@gene.com}

Authors:
\itemize{
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}
  \item Ondrej Slama \email{ondrej.slama@roche.com}
  \item Adrian Waddell \email{adrian.waddell@gene.com}
  \item Wenyi Liu \email{liu.wenyi@roche.com}
  \item Jeffrey Tomlinson \email{tomlinson.jeffrey@roche.com}
  \item Bali Toth \email{toth.balazs@gene.com}
  \item Junlue Zhao
  \item Maciej Nasinski
  \item Konrad Pagacz
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
