% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_identify_loq_values}
\alias{h_identify_loq_values}
\title{Helper for identifying any LLOQ and ULOQ values in LBSTRESC. Outcome drives
horizontal line functionality display and legend labeling along with display
of values in footnote.}
\usage{
h_identify_loq_values(loqs_data)
}
\arguments{
\item{loqs_data}{(\verb{data frame})\cr loqs_data data set containing assay data with potential LOQ values}
}
\description{
Helper for identifying any LLOQ and ULOQ values in LBSTRESC. Outcome drives
horizontal line functionality display and legend labeling along with display
of values in footnote.
}
\details{
Biomarker Sciences would like to have LLOQ and ULOQ values available for
reference in the visualizations. This also aids in setting the data constraint
ranges when goshawk functions are run from teal.goshawk UI.
}
\examples{

goshawk:::h_identify_loq_values(loqs_data = goshawk::rADLB)
}
\keyword{internal}
