% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarytable.R
\name{t_summarytable}
\alias{t_summarytable}
\title{Function to generate a table of descriptive summary statistics accompany
plots.}
\usage{
t_summarytable(data, trt_group, biomarker_var, biomarker, value_var, visit, loq,
  ...)
}
\arguments{
\item{data}{frame name}

\item{trt_group}{treatment group variable name e.g. ARM}

\item{biomarker}{PARAM value}

\item{loq}{loq variable name e.g. loq_flag (not in current ALB)}

\item{time}{visit variable name e.g. VISIT}
}
\value{
\code{tbl} object
}
\description{
Output render by teal.goshawk module \code{t_summarytable} returns
descriptive summary statistics table
}
\details{
provide additional information as needed. link to specification file \url{http://rstudio.com}
}
\examples{

ALB <- read_bce("/opt/bee/home_nas/npaszty/btk/lupus/dataadam/alb3arm.sas7bdat")
biomarker <- c('IGG') # FOR TESTING: woud come from teal.goshawk.tm_g_moduleName.R

# call function
t_summarytable(data = ALB,
               trt_group = 'ARM',
               biomarker_var = 'PARAMCD',
               biomarker = biomarker, # the PARAMCD value
               value_var = 'AVAL',
               visit = 'AVISIT',
               loq = NULL)
              

}
\author{
Balazs Toth

Nick Paszty
}
