% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarytable.R
\name{t_summarytable}
\alias{t_summarytable}
\title{Function to generate a table of descriptive summary statistics accompany
plots.}
\usage{
t_summarytable(data, biomarker, trt_group, time, loq, ...)
}
\arguments{
\item{data}{data frame name}

\item{biomarker}{biomarker PARAM value}

\item{trt_group}{treatment group variable name e.g. arm}

\item{time}{visit variable name e.g. visit}

\item{loq}{loq variable name e.g. loq_flag}
}
\value{
\code{tbl} object
}
\description{
Output render by teal.goshawk module \code{t_summarytable} returns
descriptive summary statistics table
}
\details{
provide additional information as needed. link to specification file \url{http://rstudio.com}
}
\examples{

# create test data
biomarker <- runif(100,3,100)
biomarker.bl <- NULL
for (i in 0:24) {
 biomarker.bl <- c ( biomarker.bl, rep(biomarker[i * 4 + 1], 4))
 }
 arm <- c(rep('trt',50),rep('pbo',50))
 visit <- c(rep(c(0,4,8,12),25))
 loq_flag <- c(rep(c(rep('YES',10),rep('NO',10)),5))
 example <- data.frame(biomarker = biomarker, biomarker.bl = biomarker.bl, arm = arm, loq_flag = loq_flag, visit = visit)
 colnames(example) <- c('IGG', 'IGG.bl', 'arm','loq_flag','visit')
 unit <- 'g/L'
 timepoint <- 'screening'
 color_manual <- c('pbo' = "#1F78B4", 'trt' = "#33A02C")
 shape_manual <- c('NO' = 1, 'YES' = 2, 'NA' = 0)

# call function
t_summarytable(data = example,
              biomarker = 'IGG',
              trt_group = 'arm',
              time = 'visit',
              loq = 'loq_flag')

}
\author{
Balazs Toth

Nick Paszty
}
