% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_summarytable.R
\name{t_summarytable}
\alias{t_summarytable}
\title{Function to generate a table of descriptive summary statistics accompany
plots.}
\usage{
t_summarytable(data, trt_group, param_var, param, xaxis_var,
  visit_var = "AVISITCD", loq_flag_var = "LOQFL", ...)
}
\arguments{
\item{data}{frame name}

\item{trt_group}{treatment group variable name e.g. ARM}

\item{param_var}{name of variable containing biomarker codes e.g. PARAMCD.}

\item{param}{biomarker to visualize e.g. IGG.}

\item{xaxis_var}{name of variable containing biomarker results displayed on X-axis e.g. BASE.}

\item{visit_var}{name of variable containing visit values e.g. AVISITCD}

\item{loq_flag_var}{loq variable name e.g. loq_flag_var (not in current ALB)}
}
\value{
\code{tbl} object
}
\description{
Output render by teal.goshawk module \code{t_summarytable} returns
descriptive summary statistics table as a data frame
}
\details{
provide additional information as needed. link to specification file \url{http://rstudio.com}
}
\examples{

\dontrun{
# use post processed data created in app.R file

# call function
t_summarytable(data = ALB,
               trt_group = 'ARM',
               param_var = 'PARAMCD',
               param = param, # the PARAMCD value
               xaxis_var = 'AVAL',
               visit_var = 'AVISITCD',
               loq_flag_var = 'LOQFL')

}

}
\author{
Nick Paszty (npaszty) paszty.nicholas@gene.com

Balazs Toth (tothb2)  toth.balazs@gene.com
}
