% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_hori_line_args}
\alias{validate_hori_line_args}
\title{validate arbitrary horizontal lines}
\usage{
validate_hori_line_args(
  data,
  hline_arb = NULL,
  hline_arb_color = "red",
  hline_arb_label = NULL,
  hline_vars = NULL,
  hline_vars_colors = NULL,
  hline_vars_labels = NULL,
  vline_arb = NULL,
  vline_arb_color = "green",
  vline_arb_label = NULL
)
}
\arguments{
\item{data}{data frame with variables which will be displayed in the plot.}

\item{hline_arb}{numeric value identifying intercept for arbitrary horizontal line.}

\item{hline_arb_color}{color for hline_arb that will appear on the plot.}

\item{hline_arb_label}{label for hline_arb that will appear on the legend.}

\item{hline_vars}{name(s) of variables `(ANR*)` or values `(*LOQ)` identifying intercept values.}

\item{hline_vars_colors}{color(s) for the lines of hline_arb that will appear on the plot.}

\item{hline_vars_labels}{labels(s) for hline_arb that will appear on the legend.}

\item{vline_arb}{numeric value identifying intercept for arbitrary vertical line.}

\item{vline_arb_color}{color for vline_arb that will appear on the plot.}

\item{vline_arb_label}{label for vline_arb that will appear on the legend.}
}
\description{
validate arbitrary horizontal lines
}
