% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_hori_line_args}
\alias{validate_hori_line_args}
\title{validate vertical line arguments given in the parameters against the data.}
\usage{
validate_hori_line_args(
  data,
  hline_arb = NULL,
  hline_arb_color = "red",
  hline_arb_label = NULL,
  hline_vars = NULL,
  hline_vars_colors = NULL,
  hline_vars_labels = NULL
)
}
\arguments{
\item{data}{('data.frame') with variables which will be used to create the plot.}

\item{hline_arb}{('numeric vector') value identifying intercept for arbitrary horizontal lines.}

\item{hline_arb_color}{('character vector') optional, color for the arbitrary horizontal lines.}

\item{hline_arb_label}{('character vector') optional, label for the legend to the arbitrary horizontal lines.}

\item{hline_vars}{('character vector'), names of variables `(ANR*)` or values `(*LOQ)` identifying intercept values.
The data must also contain the columns with these variable names}

\item{hline_vars_colors}{('character vector') colors for the horizontal lines defined by variables.}

\item{hline_vars_labels}{('character vector') labels for the legend to the horizontal lines defined by variables.}
}
\value{
('list')
}
\description{
helper function to be called by add_straight_lines
}
