% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_line_args}
\alias{validate_line_args}
\title{validate line arguments given in the parameters against the data.}
\usage{
validate_line_args(
  data,
  line_arb = numeric(0),
  line_arb_color = "red",
  line_arb_label = "Arbitrary line",
  line_vars = character(0),
  line_vars_colors = "green",
  line_vars_labels = line_vars
)
}
\arguments{
\item{data}{('data.frame') with variables which will be used to create the plot.}

\item{line_arb}{('numeric vector') value identifying intercept for arbitrary lines.}

\item{line_arb_color}{('character vector') optional, color for the arbitrary lines.}

\item{line_arb_label}{('character vector') optional, label for the legend to the arbitrary lines.}

\item{line_vars}{('character vector'), names of variables `(ANR*)` or values `(*LOQ)` identifying intercept values.
The data must also contain the columns with these variable names}

\item{line_vars_colors}{('character vector') colors for the lines defined by variables.}

\item{line_vars_labels}{('character vector') labels for the legend to the lines defined by variables.}
}
\value{
('list') containing the `values`,`colors` and `labels` fields defining attributes
for horizontal or vertical lines.
}
\description{
helper function to be called by add_straight_lines
}
\keyword{internal}
