% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_axes_line.R
\name{validate_line_args}
\alias{validate_line_args}
\title{Validate line arguments given in the parameters against the data.}
\usage{
validate_line_args(data, vars = character(0), color = "green", label = vars)
}
\arguments{
\item{data}{(`data.frame`)\cr
should contain `vars` which will be used to create the plot.}

\item{vars}{(`character`)\cr
names of variables `(ANR*)` or values `(*LOQ)` identifying intercept values.}

\item{color}{(`character`)\cr
colors for the lines defined by variables.}

\item{label}{(`character`)\cr
labels for the legend to the lines defined by variables.}
}
\value{
(`data.frame`) containing the `values`,`colors` and `labels` fields defining attributes
 for horizontal or vertical lines.
}
\description{
helper function to be called by [geom_axes_lines()]
}
\keyword{internal}
