% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_vert_line_args}
\alias{validate_vert_line_args}
\title{validate vertical line arguments given in the parameters against the data.}
\usage{
validate_vert_line_args(
  data,
  vline_arb = NULL,
  vline_arb_color = "red",
  vline_arb_label = NULL,
  vline_vars = NULL,
  vline_vars_colors = NULL,
  vline_vars_labels = NULL
)
}
\arguments{
\item{data}{('data.frame') with variables which will be used to create the plot.}

\item{vline_arb}{('numeric vector') value identifying intercept for arbitrary vertical lines.}

\item{vline_arb_color}{('character vector') optional, color for the arbitrary vertical lines.}

\item{vline_arb_label}{('character vector') optional, label for the legend to the arbitrary vertical lines.}

\item{vline_vars}{('character vector'), names of variables `(ANR*)` or values `(*LOQ)` identifying intercept values.
The data must also contain the columns with these variable names}

\item{vline_vars_colors}{('character vector') colors for the vertical lines defined by variables.}

\item{vline_vars_labels}{('character vector') labels for the legend to the vertical lines defined by variables.}
}
\value{
('list')
}
\description{
helper function to be called by add_straight_lines
}
